Move everything into gatsby-node and get rid of plugin
