+++ gatsby-config.js
@@ -7,7 +7,6 @@
 module.exports = {
   plugins: [
     `gatsby-plugin-sass`,
-    `source-api-plugin`,
     `gatsby-transformer-json`,
     {
       resolve: "gatsby-source-filesystem",
