--- plugins/source-api-plugin/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "source-api-plugin",
-  "version": "1.0.0",
-  "description": "Source plugin for fetching data from the api"
-}
